moved the fetch current user data api call into the MyInfo.js component. next i'll make the current user's bearer token and password a global variable to optimize the api calls.
