+++ client/ResQue_Frontend/ios/ResQue_Frontend.xcworkspace/contents.xcworkspacedata
@@ -5,9 +5,6 @@
       location = "group:../../../../../../../Downloads/gpxgenerator_path.gpx">
    </FileRef>
    <FileRef
-      location = "group:../../../../../../../Downloads/new.gpx">
-   </FileRef>
-   <FileRef
       location = "group:ResQue_Frontend.xcodeproj">
    </FileRef>
    <FileRef
