+++ client/ResQue_Frontend/screens/myinfo/MyInfo.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import { View, Text, TouchableOpacity, StyleSheet } from 'react-native';
 import IconAnt from 'react-native-vector-icons/AntDesign';
 import IconMat from 'react-native-vector-icons/MaterialIcons';
@@ -32,11 +32,75 @@ const MyinfoStack = () => {
 };
 
 const MyInfo = () => {
+
+  const [bearerToken, setBearerToken] = useState();
+  const [userData, setUserData] = useState({});
+
+  const obtainBearerToken = () => {
+    fetch('https://app-57vwexmexq-uc.a.run.app/api/users/login', {
+        method: 'POST',
+        headers: {
+            'Content-Type': 'application/json; charset=utf-8',
+        }, 
+        // Use the 'body' property to send data as JSON
+        body: JSON.stringify({
+            "email": "hello@hii.com",
+            "password": "999999999"
+        })
+    })
+    .then((response) => response.text())
+    .then((data) => {
+        if (data) {
+            console.log("there is data from the login API! it is: ", data);
+            setBearerToken(JSON.parse(data).accessToken);
+        } else {
+            console.log("there is no data from the login API...");
+        }
+    })
+    .catch((error) => {
+        console.log("error from the login API is: ", error);
+    })
+  }
+
+  const retrieveCurrentUserData = () => {
+
+      if (bearerToken) {
+          fetch('https://app-57vwexmexq-uc.a.run.app/api/users/current', {
+              method: 'GET',
+              headers: {
+                  'Content-Type': 'application/json; charset=utf-8',
+                  Authorization: `Bearer ${bearerToken}`,
+              },
+          })
+          .then((response) => response.json())
+          .then((data) => {
+              if (data) {
+                  console.log("there is data from the current user API! it is: ", data);
+                  setUserData(data);
+              } else {
+                  console.log("there is no data from the current user API...");
+              }
+          })
+          .catch((error) => {
+              console.log("error from the current user API fetch is: ", error);
+          });
+      }
+  }
+
+  useEffect( () => {
+    obtainBearerToken();
+  }, []);
+
+  useEffect( () => {
+      retrieveCurrentUserData()
+  }, [bearerToken]);
+
+
   return (
     <View style={styles.container}>
       <View style={styles.user}>
         <IconMat name="circle" size={90} color="#CC313D" />
-        <Text style={styles.userName}>Jane Doe</Text>
+        <Text style={styles.userName}>{userData.firstName} {userData.lastName}</Text>
       </View>
       <View style={styles.infoContainer}>
         <InfoItem icon="person-outline" text="Edit profile" screen="EditProfile" />
